pyproject.toml
[project]
name = "argus"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.58",
    "psycopg[binary]>=3.3",
    "apscheduler>=3.11",
    "python-dotenv>=1.2",
]

[project.optional-dependencies]
dev = ["pytest==9.0.3"]
